pyproject.toml
[project]
name = "claude-to-im"
version = "0.1.0"
description = "Claude Code CLI to IM bridge service"
requires-python = ">=3.11"
dependencies = [
    "lark-oapi>=1.2.0",
    "claude-agent-sdk>=0.1.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
